Commit pada 22 Desember 2023 pukul 23:26 WIB

+++ app/src/main/java/com/example/uas/MovieAddActivity.kt
@@ -105,7 +105,7 @@ class MovieAddActivity : AppCompatActivity() {
                                             id = newDocumentId.toString(),
                                             gambar = filePath, // Menggunakan jalur file lokal
                                             nama = inputMovieTitle.text.toString(),
-                                            rating = ratingSlider.value.toDouble(),
+                                            rating = ratingSlider.value.toInt(),
                                             direktor = inputMovieDir.text.toString(),
                                             genre = mMovieDao.fromListToString(selectedGenres).toString(),
                                             storyline = inputStoryLine.text.toString()
